custom exception handling

+++ src/main/java/com/arslanca/dev/business/concretes/PinnedProjectManager.java
@@ -5,6 +5,8 @@ import com.arslanca.dev.business.dto.requests.CreatePinnedProjectRequest;
 import com.arslanca.dev.business.dto.requests.UpdatePinnedProjectRequest;
 import com.arslanca.dev.business.dto.responses.PinnedProjectResponse;
 import com.arslanca.dev.business.mappers.PinnedProjectMapper;
+import com.arslanca.dev.core.utilities.exceptions.types.BusinessException;
+import com.arslanca.dev.core.utilities.exceptions.types.NotFoundException;
 import com.arslanca.dev.dataAccess.PinnedProjectRepository;
 import com.arslanca.dev.entities.PinnedProject;
 import lombok.RequiredArgsConstructor;
@@ -28,19 +30,32 @@ public class PinnedProjectManager implements PinnedProjectService {
 
     @Override
     public void add(CreatePinnedProjectRequest request) {
+        if (repository.existsByTitle(request.getTitle())) {
+            throw new BusinessException("Bu başlıkta bir proje zaten mevcut: " + request.getTitle());
+        }
         PinnedProject project = mapper.toEntity(request);
         repository.save(project);
     }
 
     @Override
     public void update(Long id, UpdatePinnedProjectRequest request) {
-        PinnedProject project = repository.findById(id).orElseThrow();
+        PinnedProject project = repository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Proje bulunamadı (ID: " + id + ")"));
+
+        // Title check if title is changed
+        if (!project.getTitle().equalsIgnoreCase(request.getTitle()) && repository.existsByTitle(request.getTitle())) {
+            throw new BusinessException("Girdiğiniz yeni başlık başka bir projede kullanılıyor: " + request.getTitle());
+        }
+
         mapper.updateEntityFromRequest(request, project);
         repository.save(project);
     }
 
     @Override
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new NotFoundException("Silinecek proje bulunamadı (ID: " + id + ")");
+        }
         repository.deleteById(id);
     }
 }
